+++ Adj/Algebra/Set.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE EmptyCase #-}
+
 module Adj.Algebra.Set where
 
 infixr 7 :*:, :+:
@@ -7,3 +9,17 @@ data (:*:) l r = l :*: r
 
 -- Disjoint union
 data (:+:) l r = This l | That r
+
+data Unit = Unit
+
+boring :: o -> Unit
+boring _ = Unit
+
+data Void
+
+absurd :: Void -> o
+absurd x = case x of {}
+
+type family Neutral p = r | r -> p where
+	Neutral (:*:) = Unit
+	Neutral (:+:) = Void